chore: move Ext2Intt event to stdlib
