+++ assembly/src/parser/grammar.lalrpop
@@ -49,7 +49,6 @@ extern {
         "adv_loadw" => Token::AdvLoadw,
         "adv_pipe" => Token::AdvPipe,
         "adv_push" => Token::AdvPush,
-        "push_ext2intt" => Token::PushExt2intt,
         "push_mapval" => Token::PushMapval,
         "push_mapvaln" => Token::PushMapvaln,
         "push_mtnode" => Token::PushMtnode,
@@ -668,7 +667,6 @@ SystemEvent: Instruction = {
     "adv" "." "insert_hdword_d" => Instruction::SysEvent(SystemEventNode::InsertHdwordWithDomain),
     "adv" "." "insert_hperm" => Instruction::SysEvent(SystemEventNode::InsertHperm),
     "adv" "." "insert_mem" => Instruction::SysEvent(SystemEventNode::InsertMem),
-    "adv" "." "push_ext2intt" => Instruction::SysEvent(SystemEventNode::PushExt2intt),
     "adv" "." "push_mapval" => Instruction::SysEvent(SystemEventNode::PushMapVal),
     "adv" "." "push_mapvaln" => Instruction::SysEvent(SystemEventNode::PushMapValN),
     "adv" "." "push_mtnode" => Instruction::SysEvent(SystemEventNode::PushMtNode),
