+++ assembly/src/parser/token.rs
@@ -144,7 +144,6 @@ pub enum Token<'input> {
     AdvLoadw,
     AdvPipe,
     AdvPush,
-    PushExt2intt,
     PushMapval,
     PushMapvaln,
     PushMtnode,
@@ -327,7 +326,6 @@ impl fmt::Display for Token<'_> {
             Token::AdvLoadw => write!(f, "adv_loadw"),
             Token::AdvPipe => write!(f, "adv_pipe"),
             Token::AdvPush => write!(f, "adv_push"),
-            Token::PushExt2intt => write!(f, "push_ext2intt"),
             Token::PushMapval => write!(f, "push_mapval"),
             Token::PushMapvaln => write!(f, "push_mapvaln"),
             Token::PushMtnode => write!(f, "push_mtnode"),
@@ -518,7 +516,6 @@ impl<'input> Token<'input> {
                 | Token::AdvLoadw
                 | Token::AdvPipe
                 | Token::AdvPush
-                | Token::PushExt2intt
                 | Token::PushMapval
                 | Token::PushMapvaln
                 | Token::PushMtnode
@@ -661,7 +658,6 @@ impl<'input> Token<'input> {
         ("adv_loadw", Token::AdvLoadw),
         ("adv_pipe", Token::AdvPipe),
         ("adv_push", Token::AdvPush),
-        ("push_ext2intt", Token::PushExt2intt),
         ("push_mapval", Token::PushMapval),
         ("push_mapvaln", Token::PushMapvaln),
         ("push_mtnode", Token::PushMtnode),
