+++ processor/src/lib.rs
@@ -61,7 +61,7 @@ use trace::TraceFragment;
 pub use trace::{ChipletsLengths, ExecutionTrace, TraceLenSummary, NUM_RAND_ROWS};
 
 mod errors;
-pub use errors::{ExecutionError, Ext2InttError};
+pub use errors::ExecutionError;
 
 pub mod utils;
